LayoutTests/platform/mac-wk1/fast/repaint/drop-shadow-box-change-expected.txt

Test repaint rects when a box with drop-shadow changes size

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL layerTreeAsText.indexOf('(rect 192 192 206 156)') > -1 should be true. Was false.
PASS successfullyParsed is true

TEST COMPLETE
